requirements.txt
# وابستگی‌های ربات تلگرام مدیریت پنل 3X-UI

# کتابخانه تلگرام
python-telegram-bot==20.7

# کتابخانه HTTP غیرهمزمان
aiohttp==3.9.1

# زمان‌بندی تسک‌ها
apscheduler==3.10.4

# دیتابیس SQLite غیرهمزمان
aiosqlite==0.19.0
